Add files via upload
I completely redid the UI and UX. Its much more organized now. I used the Bootstrap Icon Library to import the icons into our menus. I used the ToopTip feature to add ToolTips to all of the buttons. The tooltip allows you to display a small text box with information when hovering over a button. In this case it details whats inside of the buttons.